sessions: fix type errors when calling nox.session
This worked previously because of the way the actual function is
defined, but that's an implementation detail. The typing overloads state
that this should be called like a function decorator — either plain
`@nox.session` (equivalent to `nox.session(func)`) or with arguments,
`@nox.section(name=...)` (equivalent to `nox.session(name=...)(func)`).