adjust the ee templates and generator
The resulting ee yaml file contained a bunch of escaped strings so I
replaced store_yaml_file() to just write the file directly. Also made
some changes to the jinja template and matrix config after noticing that
the structure was not as expected and did not align with the
ansible-builder docs.