Allow to install packages in editable mode (#106)
* Allow to install packages in editable mode.

* Draft: propose another mechanism for specifying packages

This adds support for specifying packages as a dictionary with a type in
addition to a simple string. This allows more flexibility to specify
other types of dependencies and requires less special casing and string
parsing than the previous solution.

* utils: remove special handling of -e in install()

This isn't necessary anymore, as nothing else should directly pass "-e"
to the install() function.

* Draft: Comment out PackageConsraints pending feedback

Unless the `_package` options are modified to accept a list of multiple
PackageTypes, this `PackageConsraints` should just be removed entirely.

* Separate config (using Pydantic) and runtime (using dataclass) types.

---------

Co-authored-by: Maxwell G <maxwell@gtmx.me>